Better log info in pagination warning
The Steam Card Exchange can list apps that have items which can no longer be purchased, and querying the Steam API for these apps will not provide details on those items, even though they are still counted as part of the result count in a response payload. For now, just print out the URL and manually check if there aren't any items listed.